# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "W&M Store",
        "description": "Listado de URI's of Agency API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:85"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List of all Customers",
                "operationId": "2b23ac696cfd81c23807c2c1fd63b0b0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "agency_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "dni": {
                                                        "type": "string",
                                                        "example": "012345678"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Anibal Gallardo"
                                                    },
                                                    "phone_numbers": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "phone": {
                                                                    "type": "string",
                                                                    "example": "012345678"
                                                                },
                                                                "operator_id": {
                                                                    "type": "number",
                                                                    "example": 1
                                                                },
                                                                "operator": {
                                                                    "type": "string",
                                                                    "example": "Entel"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Register Customer",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "agency_id": {
                                                "type": "number"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "dni": {
                                                "type": "string"
                                            },
                                            "phone_numbers": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "012345678"
                                                        },
                                                        "operator_id": {
                                                            "type": "number",
                                                            "example": 1
                                                        },
                                                        "operator": {
                                                            "type": "string",
                                                            "example": "Entel"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "operator_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Anibal Gallardo",
                                    "address": "Tacna Peru",
                                    "phone": "906393152",
                                    "email": "anibal.gg.206@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Error Object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/custormers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Show one customer",
                "operationId": "f91d42f280dce9f47bb649baa6703720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "agency_id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "dni": {
                                            "type": "string",
                                            "example": "01234567"
                                        },
                                        "phone_numbers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "012345678"
                                                    },
                                                    "operator_id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "operator": {
                                                        "type": "string",
                                                        "example": "Entel"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/customers/{id}": {
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update agency",
                "operationId": "07a3f33c1373d3c6a97b7d8846d81472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "agency_id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "dni": {
                                                "type": "string"
                                            },
                                            "operator_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "agency_id": 1,
                                    "name": "Anibal Gallardo Gonzalez",
                                    "dni": "01234567",
                                    "operator_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Error Object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete Customer",
                "operationId": "634844e24d72540f608f7fd297f5d8b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "Enter your Bearer token in the format **Bearer {token}**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}